+++ crates/searcher/src/searcher/core.rs
@@ -612,6 +612,17 @@ impl<'s, M: Matcher, S: Sink> Core<'s, M, S> {
             return false;
         }
         if let Some(line_term) = self.matcher.line_terminator() {
+            // FIXME: This works around a bug in grep-regex where it does
+            // not set the line terminator of the regex itself, and thus
+            // line anchors like `(?m:^)` and `(?m:$)` will not match
+            // anything except for `\n`. So for now, we just disable the fast
+            // line-by-line searcher which requires the regex to be able to
+            // deal with line terminators correctly. The slow line-by-line
+            // searcher strips line terminators and thus absolves the regex
+            // engine from needing to care about whether they are `\n` or NUL.
+            if line_term.as_byte() == b'\x00' {
+                return false;
+            }
             if line_term == self.config.line_term {
                 return true;
             }